+++ projects/batfish-common-protocol/src/main/java/org/batfish/specifier/IpSpaceSpecifierFactory.java
@@ -1,11 +1,23 @@
 package org.batfish.specifier;
 
+import java.util.ServiceLoader;
+import org.batfish.common.BatfishException;
+
 /**
  * Used by {@link org.batfish.datamodel.questions.Question}s to build {@link IpSpaceSpecifier}
  * objects. The factory can be responsible for parsing and validating the fields from question
  * input.
  */
 public interface IpSpaceSpecifierFactory {
+  static IpSpaceSpecifierFactory load(String name) {
+    for (IpSpaceSpecifierFactory factory : ServiceLoader.load(IpSpaceSpecifierFactory.class)) {
+      if (factory.getName().equals(name)) {
+        return factory;
+      }
+    }
+    throw new BatfishException("Could not find an IpSpaceSpecifierFactory with name " + name);
+  }
+
   String getName();
 
   /**
